Try using linux homebrew for gcm

+++ post_install.sh
@@ -30,20 +30,17 @@ if os_is mac; then
 	os_install mono-libgdiplus
 fi
 # https://learn.microsoft.com/en-us/dotnet/core/tools/dotnet-install-script
-wget https://dot.net/v1/dotnet-install.sh
-chmod +x dotnet-install.sh
-./dotnet-install.sh --channel 7.0
-rm dotnet-install.sh
+if have_not_installed dotnet; then
+	wget https://dot.net/v1/dotnet-install.sh
+	chmod +x dotnet-install.sh
+	./dotnet-install.sh --channel 7.0
+	rm dotnet-install.sh
+fi
 
 # Git credential manager
 # https://github.com/git-ecosystem/git-credential-manager/blob/release/docs/install.md
 if have_not_installed git-credential-manager; then
-	if os_is mac; then
-		brew install --cask git-credential-manager
-	fi
-	if os_is ubuntu; then
-		dotnet tool install -g git-credential-manager
-	fi
+	brew install --cask git-credential-manager
 fi
 
 # zsh-abbr
